move mconfig_digest.go functions to the cloud
